+++ .gitignore
@@ -1,3 +1,4 @@
 env/
 khmer/
 setup
+nullgraph/
